+++ DatadogLogs/Sources/Log/LogEventBuilder.swift
@@ -75,7 +75,12 @@ internal struct LogEventBuilder {
             buildId: context.buildId,
             variant: context.variant,
             dd: LogEvent.Dd(
-                device: LogEvent.DeviceInfo(architecture: context.device.architecture)
+                device: LogEvent.DeviceInfo(
+                    brand: context.device.brand,
+                    name: context.device.name,
+                    model: context.device.model,
+                    architecture: context.device.architecture
+                )
             ),
             os: .init(
                 name: context.device.osName,